# Conflicts:
#	trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/MatrixClient.kt
#	trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/key/KeyService.kt
#	trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/room/RoomService.kt
#	trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/user/UserService.kt
#	trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/verification/VerificationService.kt
#	trixnity-clientserverapi/trixnity-clientserverapi-client/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/client/SyncApiClient.kt

+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/RoomDisplayName.kt
@@ -1,14 +1,14 @@
 package net.folivo.trixnity.client.store
 
 import kotlinx.serialization.Serializable
-import net.folivo.trixnity.clientserverapi.model.sync.SyncResponse.Rooms.JoinedRoom.RoomSummary
+import net.folivo.trixnity.clientserverapi.model.sync.Sync.Response.Rooms.JoinedRoom.RoomSummary
 import net.folivo.trixnity.core.model.UserId
 
 @Serializable
 data class RoomDisplayName(
     val explicitName: String? = null,
     val isEmpty: Boolean = false,
-    val otherUsersCount: Int = 0,
+    val otherUsersCount: Long = 0,
     internal val summary: RoomSummary?
 ) {
     val heroes: List<UserId>
